dipper: overlay-lineage: Allow fps listening when screen is off
The overlay used to hide unnecessary preference for this target, despite
the feature was still supported under-the-hood. Moreover, the shipped
PocketMode app checked for proximity sensor to decide whether to allow
fingerprint auth to work.

The logic changed recently and now the preference must be visible so
that users can actually select the desired behaviour, since listening
is disabled by default.

Change-Id: Ib22af7b6e48584bc51dc40b5572af149410f476a